move open ai tools to chat_completions.py
